Commands "gg" and "G" not changing horizontal pos
It was also added a small helper function to process events
with a QEventLoop for a maximum timeout. This was done because
since the QTRY_* macros only process events *between* comparisons,
some tests had bugs because events should be processed before
the first comparison, e.g., the initial X position is 100 and
the "gg" command mistakenly changed it to 0. The test was not
caughting this bug because in the first comparison X still have
its old value.

Closes #1