Added fixed size scroll
Before the scroll was starting as soon as the user pressed
the button and finishing as soon as he releases. This made
the scroll size a bit unpredictable even for single keystrokes
because the interval between pressing and release varies a
lot in miliseconds world. Changed the scroll to perform at
least 10 individual steps in each scroll. It seems much more
predictable now. Vimium in Chromium is predictable as well.

It was also changed the size of the half page scroll. Instead
of scrolling half window size in each step it's now scrolling
window_size/10 in order to really scroll to middle of viewport.