Removed boilerplate code from tests
A lot of tests were repeating the same code to load the test
page. Moved this to init() which is called before every test.

It was added cleanup() to close the current tab in order to
avoid dependencies as much as possible among tests.

Set test page as Qt resource for portability. The test is now
moving the page to "tmp" and deleting the file after in order
to QUrl have access to the resource.