Fixed current view in some tests
BrowserWindow has a list of WebViews. When tab tests got
a reference of the current WebView, this reference was
stricly bound to the current tab at the moment, ie,
BrowserWindow::weView() returns a different object when
current tab changes. Despite the fact tests were not failing
because of this, it was a flaw in the tests.