Sending events to 'focusProxy' instead of 'parentWidget'
Events are forwarded to parentWidget if they were not consumed
by the page, ie, if the page does something with a given key
then the keyPress event is not forward to parentWidget. Since
in the tests we were sending the events directly to parentWidget,
we missed the bug in qt-webengine 5.9 - unconsumed events not
being forwarded to parentWidget. This commit changes the tests
to send the event to 'focusProxy' likewise. As a side effect
the tests work on qt-webengine 5.8 but DON'T on 5.9. This is
the expected behaviour until events start being forward again.

Issue #13